move PixelSender interface to the api module
